Added a working call to create user

+++ bank-api/src/main/java/com/bank/bankapi/resources/EmployeeResources.java
@@ -2,6 +2,7 @@ package com.bank.bankapi.resources;
 
 import com.bank.bankapi.Constants;
 import com.bank.bankapi.domain.Employee;
+import com.bank.bankapi.repositories.EmployeeRepository;
 import com.bank.bankapi.services.EmployeeService;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
@@ -23,16 +25,27 @@ public class EmployeeResources {
 
     @Autowired
     EmployeeService employeeService;
+    @Autowired
+    EmployeeRepository employeeRepository;
 
     @PostMapping("/register")
-    public ResponseEntity<Map<String,String>> registerEmployee(@RequestBody Map<String,Object> data){
+    public ResponseEntity<Map<String,String>> registerEmployee(@RequestBody Map<String,Object> data, HttpServletRequest request){
+        int userID= (Integer) request.getAttribute("userId");
+        Employee employeeAuth =  employeeRepository.findEmployeeById(userID);
+        if(employeeAuth==null ||employeeAuth.getRole() != Employee.Role.ADMIN )
+        {
+            Map<String, String> map= new HashMap<>();
+            map.put("message","You are not authorized to do this function");
+            return new ResponseEntity<>(map,HttpStatus.OK);
+        }
         String firstName = (String)data.get("firstName");
         String lastName= (String)data.get("lastName");
         String phone= (String)data.get("phone");
         String password=(String)data.get("password");
         Employee employee= employeeService.registerEmployee(firstName,lastName,password, Employee.Role.EMPLOYEE,phone);
-
-        return new ResponseEntity<>(generateJWTToken(employee), HttpStatus.OK);
+        Map<String ,String > map = new HashMap<>();
+        map.put("message","Employee has been registered");
+        return new ResponseEntity<>(map, HttpStatus.OK);
 
     }
 
@@ -48,11 +61,17 @@ public class EmployeeResources {
     }
 
     @PostMapping("/delete")
-    public ResponseEntity<Map<String,String>> deleteEmployee(@RequestBody Map<String,Object> data){
-
+    public ResponseEntity<Map<String,String>> deleteEmployee(@RequestBody Map<String,Object> data, HttpServletRequest request){
+        int userID= (Integer) request.getAttribute("userId");
+        Employee employeeAuth =  employeeRepository.findEmployeeById(userID);
+        Map<String, String> map = new HashMap<>();
+        if(employeeAuth==null ||employeeAuth.getRole() != Employee.Role.ADMIN )
+        {
+            map.put("message","You are not authorized to do this function");
+            return new ResponseEntity<>(map,HttpStatus.OK);
+        }
         String phone= (String)data.get("phone");
         boolean flag= employeeService.deleteEmployee(phone);
-        Map<String, String> map = new HashMap<>();
         if(flag)
             map.put("message","Employee Deleted Successfully");
         else
